Move github models to github folder
We initially placed all models in the models folder. Then we started
grouping them by functionality. This change moves all remaining
GitHub-related models to the github folder.
