+++ model/github/github_runner.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative "../model"
-require_relative "../lib/net_ssh"
+require_relative "../../model"
+require_relative "../../lib/net_ssh"
 
 class GithubRunner < Sequel::Model
   one_to_one :strand, key: :id
